# backend/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Geevheit Intelligence 360 API
  version: 0.1.0
  description: API-first contract for the enterprise transportation intelligence platform.
paths:
...
  /api/v1/financial/transactions:
    get:
      summary: Search financial transactions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Financial transaction page
  /api/v1/financial/transactions/{id}:
    get:
      summary: Get financial transaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Financial transaction
    put:
      summary: Update draft financial transaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Financial transaction updated
  /api/v1/financial/transactions/{id}/approve:
    post:
      summary: Approve financial transaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Financial transaction approved
  /api/v1/financial/transactions/{id}/pay:
    post:
      summary: Pay expense transaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Expense paid
  /api/v1/financial/transactions/{id}/receive:
    post:
      summary: Receive revenue transaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Revenue received
  /api/v1/financial/transactions/{id}/adjust:
    post:
      summary: Adjust finalized financial transaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "201":
          description: Financial adjustment created
  /api/v1/financial/expenses:
    get:
      summary: Search expenses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Expense page
    post:
      summary: Create expense
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Expense created
  /api/v1/financial/revenues:
    get:
      summary: Search revenues
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Revenue page
    post:
      summary: Create revenue
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Revenue created
  /api/v1/financial/categories:
    get:
      summary: Search cost categories
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Category page
    post:
      summary: Create cost category
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Category created
  /api/v1/financial/types:
    get:
      summary: Search cost types
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cost type page
    post:
      summary: Create cost type
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Cost type created
  /api/v1/financial/cost-centers:
    get:
      summary: Search cost centers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cost center page
    post:
      summary: Create cost center
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Cost center created
  /api/v1/financial/accounts:
    get:
      summary: Search financial accounts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account page
    post:
      summary: Create financial account
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Account created
  /api/v1/financial/payment-methods:
    get:
      summary: Search payment methods
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Payment method page
    post:
      summary: Create payment method
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Payment method created
  /api/v1/financial/periods:
    get:
      summary: Search financial periods
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Financial period page
    post:
      summary: Create financial period
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Financial period created
  /api/v1/financial/periods/{id}/close:
    post:
      summary: Close financial period
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Financial period closed
  /api/v1/financial/budgets:
    get:
      summary: Search budgets
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Budget page
    post:
      summary: Create budget
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Budget created
  /api/v1/financial/adjustments:
    get:
      summary: Search financial adjustments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Financial adjustment page
    post:
      summary: Adjust finalized financial transaction
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Financial adjustment created
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
...